requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
Django==4.2.6
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
whitenoise==6.6.0
